so far so good

+++ src/Graphics/color.h
@@ -5,9 +5,9 @@
 
 typedef struct
 {
-  uint16_t red;
-  uint16_t green;
-  uint16_t blue;
+  uint8_t red;
+  uint8_t green;
+  uint8_t blue;
 }
 Color;
